backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.30
apscheduler==3.10.4
requests==2.32.3
python-dotenv==1.0.1
firebase-admin==6.5.0
protobuf==4.25.3
gtfs-realtime-bindings==1.0.0
pytest==8.2.2
pytest-cov==5.0.0
